LayoutTests/media/modern-media-controls/ios-inline-media-controls/touch/ios-inline-media-controls-shows-start-button-expected.txt

Check that tapping the start button actually plays the video.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS shadowRoot.querySelector('button.play-pause') became different from null
PASS shadowRoot.querySelector('button.play-pause').getBoundingClientRect().width became different from 0
Tapping start button...
Waiting for playback...
PASS media.paused became false
Checking again...
PASS media.paused is false
PASS successfullyParsed is true

TEST COMPLETE
